Refactor chronostratigraphyEditProfile
