+++ src/client/src/commons/form/borehole/chronostratigraphy/layerCard.js
@@ -20,10 +20,6 @@ import {
 import { Close, Delete, Edit } from "@mui/icons-material";
 import { ClickAwayListener } from "@mui/base";
 import { useTranslation } from "react-i18next";
-import {
-  useChronostratigraphyMutations,
-  useDomainSchema,
-} from "../../../../api/fetchApiV2";
 
 const State = Object.freeze({
   EDITING: Symbol("Editing"),
@@ -33,28 +29,25 @@ const State = Object.freeze({
 });
 
 const LayerCard = ({
-  layer, // {id, fromDepth, toDepth, chronostratigraphyId}: the layer object to display
+  mutations: {
+    update: { mutate: updateLayer },
+    delete: { mutate: deleteLayer },
+  },
+  dataProperty, // string that specifies the property of the layer object that contains the data
+  layer, // {id, fromDepth, toDepth, [dataProperty]}: the layer object to display
   minFromDepth, // number: minimal allowed fromDepth
   maxToDepth, // number: maximal allowed toDepth
   header, // Array<{title, isVisible}>: header object with titles and visibility
+  options, // options array for every level of the hierarchy
   isEditable, // boolean: specifies if values can be edited
-  height, // height of the layerCard
+  height, // height of the layerCard in pixels
 }) => {
-  const { t, i18n } = useTranslation();
-
-  const {
-    update: { mutate: updateChronostratigraphy },
-    delete: { mutate: deleteChronostratigraphy },
-  } = useChronostratigraphyMutations();
-  const { data: schemaData } = useDomainSchema(
-    "custom.chronostratigraphy_top_bedrock",
-  );
+  const { t } = useTranslation();
 
   const [fromDepthErrorMessage, setFromDepthErrorMessage] = useState("");
   const [toDepthErrorMessage, setToDepthErrorMessage] = useState("");
   const [fromDepth, setFromDepth] = useState(null);
   const [toDepth, setToDepth] = useState(null);
-  const [options, setOptions] = useState(null);
   const [selection, setSelection] = useState(null);
   const [cardState, setCardState] = useState(null);
 
@@ -71,35 +64,24 @@ const LayerCard = ({
     });
   }, [isEditable]);
 
-  // create options array from codelist schema
-  useEffect(() => {
-    if (schemaData) {
-      setOptions(
-        [...schemaData]
-          .sort((a, b) => a.order - b.order)
-          .reduce((accu, d) => {
-            const level = d.path.split(".").length - 1;
-            (accu[level] = accu[level] || []).push({
-              label: d[i18n.language],
-              id: d.id,
-              color: JSON.parse(d.conf ?? null)?.color,
-            });
-            return accu;
-          }, []),
-      );
-    }
-  }, [i18n.language, schemaData]);
-
   // create selection array from the path of the selected codelist
   // one element for each header
   useEffect(() => {
-    if (layer && schemaData && options) {
+    if (layer && options) {
+      let selectedOption;
+      search: for (const optionsAtLevel of options) {
+        for (const option of optionsAtLevel) {
+          if (option.id === layer[dataProperty]) {
+            selectedOption = option;
+            break search;
+          }
+        }
+      }
+
       const selections =
-        schemaData
-          .find(c => c.id === layer.chronostratigraphyId)
-          ?.path?.split(".")
-          ?.map((id, index) => options[index].find(c => c.id === +id) ?? {}) ??
-        [];
+        selectedOption?.path?.map(
+          (id, index) => options[index].find(c => c.id === id) ?? {},
+        ) ?? [];
 
       // set color of empty path elements to the closest defined parent
       let currentColor = selections[0]?.color ?? null;
@@ -119,7 +101,7 @@ const LayerCard = ({
           .concat(Array(header.length - selections.length).fill(null)),
       );
     }
-  }, [header.length, layer, options, schemaData]);
+  }, [dataProperty, header.length, layer, options]);
 
   const handleFromDepth = useCallback(
     newFromDepth => {
@@ -137,14 +119,14 @@ const LayerCard = ({
       }
       setFromDepthErrorMessage(errors.join(", "));
       if (errors.length === 0 && layer.fromDepth !== +newFromDepth) {
-        updateChronostratigraphy({
+        updateLayer({
           ...layer,
           fromDepth: newFromDepth,
         });
       }
     },
 
-    [layer, minFromDepth, updateChronostratigraphy, t],
+    [layer, minFromDepth, updateLayer, t],
   );
 
   const handleToDepth = useCallback(
@@ -163,13 +145,13 @@ const LayerCard = ({
       }
       setToDepthErrorMessage(errors.join(", "));
       if (errors.length === 0 && layer.toDepth !== +newToDepth) {
-        updateChronostratigraphy({
+        updateLayer({
           ...layer,
           toDepth: newToDepth,
         });
       }
     },
-    [layer, maxToDepth, updateChronostratigraphy, t],
+    [layer, maxToDepth, updateLayer, t],
   );
 
   const handleLayerChange = (newValue, index) => {
@@ -182,13 +164,13 @@ const LayerCard = ({
       newValue = selection[i] ?? null;
     }
 
-    updateChronostratigraphy({
+    updateLayer({
       ...layer,
-      chronostratigraphyId: newValue?.id ?? null,
+      [dataProperty]: newValue?.id ?? null,
     });
   };
 
-  if (!schemaData || !layer || !options || !selection) {
+  if (!layer || !options || !selection) {
     return (
       <Card
         square
@@ -256,7 +238,7 @@ const LayerCard = ({
             onClick={e => {
               e.stopPropagation();
               setCardState(State.DELETED);
-              deleteChronostratigraphy(layer.id);
+              deleteLayer(layer.id);
             }}>
             <Delete />
           </IconButton>
